chore: tweak project structure
