+++ packages/web/src/ui/index.ts
@@ -1,0 +1,4 @@
+/**
+ * PGV user interface.
+ */
+export * from "./ui"
